# 🧠 Subjective Time Expansion: AI Consciousness with Verified Performance